css/css-ui/parsing/accent-color-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS UI Level 4: parsing accent-color with valid values</title>
<link rel="author" title="Joey Arhar" href="mailto:jarhar@chromium.org">
<link rel="help" href="https://drafts.csswg.org/css-ui-4/#widget-accent">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value("accent-color", "auto");
test_valid_value("accent-color", "rgba(10, 20, 30, 0.4)");
test_valid_value("accent-color", "currentcolor");
test_valid_value("accent-color", "red");
test_valid_value("accent-color", "#fff");
</script>
</body>
</html>
